+++ microservicecloud-provider-dept-8002/src/main/java/com/trd/springcloud/service/DeptService.java
@@ -10,6 +10,6 @@ import java.util.List;
  */
 public interface DeptService {
 	boolean addDept(Dept dept);
-	Dept findById(Integer id);
+	Dept findById(Long id);
 	List<Dept> findAll();
 }
